Added CPath class, replaced angularVelocity function
Created new CPath class.  Created it as a static class for now.  My
intention is to consolidate duplicate functionality here.  Eventually this
class may be converted to a non-static class where the collected functions
become proper methods.

I also replaced the angularVelocity function I added in my previous commit
with a new VelocityLimitSteerAngle function in CPath.  This new function
reduces the steering angle if the maximum angular velocity will be exceeded.

I then replaced the duplicate in each of five classes with a call to this
function.