Added information to message boxes
I spent a lot of time when I started with AOG trying to figure out what was accutually wrong when I got and error message box.  I added a comment to check the log file for details and added more descriptive messages to the message box that pops up when driving a recorded path fails.